extract and move selected weather component to own file and create folder for generic weathercomponents
