+++ features/weather/LocationForecast.tsx
@@ -69,7 +69,7 @@ export const WeatherComponent = ({ weatherdata }: WeatherProps) => {
 
   return (
     <View style={{ padding: 8 }}>
-      <WeatherToday
+      <SelectedWeather
         time={displayHour.time}
         symbol={displayHour.data.next_1_hours?.summary.symbol_code}
         today={displayHour.data.instant.details}
@@ -93,261 +93,6 @@ export const WeatherComponent = ({ weatherdata }: WeatherProps) => {
     </View>
   );
 };
-// ------ WEATHER NOW LOGIC ---------
-type TodayProps = {
-  time: string;
-  symbol?: string;
-  today: WeatherInstantDetails;
-};
-const WeatherToday = ({ time, symbol, today }: TodayProps) => {
-  return (
-    <View>
-      <View style={{ flexDirection: "row", gap: 10 }}>
-        <View>
-          <Text variant="titleLarge">Today {toTime(time)}</Text>
-          <WeatherDetail
-            icon="sun-thermometer"
-            value={today.air_temperature}
-            unit="°C"
-            textSize="titleMedium"
-          />
-          <WindDetail
-            speed={today.wind_speed}
-            direction={today.wind_from_direction}
-            unit="m/s"
-          />
-        </View>
-        {symbol && weatherIcons[symbol] && (
-          <Image
-            source={weatherIcons[symbol]}
-            style={{ width: 80, height: 80 }}
-          />
-        )}
-      </View>
-      <WeatherDetailsCollapsible details={today} textSize="bodyMedium" />
-    </View>
-  );
-};
-
-function degreesToCompass(deg?: number): string {
-  if (deg === undefined) return "";
-  const dirs = [
-    "N",
-    "NNE",
-    "NE",
-    "ENE",
-    "E",
-    "ESE",
-    "SE",
-    "SSE",
-    "S",
-    "SSW",
-    "SW",
-    "WSW",
-    "W",
-    "WNW",
-    "NW",
-    "NNW",
-  ];
-  const ix = Math.round((deg % 360) / 22.5) % 16;
-  return dirs[ix];
-}
-
-type WindDetailProps = {
-  speed?: number;
-  direction?: number;
-  unit?: string;
-};
-
-export const WindDetail: React.FC<WindDetailProps> = ({
-  speed,
-  direction,
-  unit = "m/s",
-}) => {
-  const { colors } = useTheme();
-  // Arrow should point to where the wind is going (from + 180°)
-  const arrowRotation = direction !== undefined ? (direction + 180) % 360 : 0;
-  return (
-    <View
-      style={{
-        flexDirection: "row",
-        alignItems: "center",
-        gap: 8,
-      }}
-    >
-      <WeatherDetail
-        icon="weather-windy"
-        value={speed}
-        unit={unit}
-        textSize="titleMedium"
-      />
-      {direction !== undefined && (
-        <>
-          <Text variant="titleMedium">
-            {`${direction}° ${degreesToCompass(direction)}`}
-          </Text>
-          <MaterialCommunityIcons
-            name="arrow-up"
-            size={20}
-            color={colors.primary}
-            style={{ transform: [{ rotate: `${arrowRotation}deg` }] }}
-          />
-        </>
-      )}
-    </View>
-  );
-};
-
-type WeatherDetailsListProps = {
-  details: WeatherInstantDetails;
-  textSize?: "titleSmall" | "titleMedium" | "bodyMedium" | "bodySmall";
-};
-
-const detailConfig = [
-  {
-    key: "air_pressure_at_sea_level",
-    label: "Pressure",
-    unit: "hPa",
-    icon: "gauge",
-  },
-  {
-    key: "air_temperature",
-    label: "Temperature",
-    unit: "°C",
-    icon: "thermometer",
-  },
-  {
-    key: "cloud_area_fraction",
-    label: "Cloud Cover",
-    unit: "%",
-    icon: "weather-cloudy",
-  },
-  {
-    key: "cloud_area_fraction_high",
-    label: "High Clouds",
-    unit: "%",
-    icon: "weather-cloudy",
-  },
-  {
-    key: "cloud_area_fraction_medium",
-    label: "Medium Clouds",
-    unit: "%",
-    icon: "weather-cloudy",
-  },
-  {
-    key: "cloud_area_fraction_low",
-    label: "Low Clouds",
-    unit: "%",
-    icon: "weather-cloudy",
-  },
-  {
-    key: "dew_point_temperature",
-    label: "Dew Point",
-    unit: "°C",
-    icon: "water",
-  },
-  { key: "fog_area_fraction", label: "Fog", unit: "%", icon: "weather-fog" },
-  {
-    key: "relative_humidity",
-    label: "Humidity",
-    unit: "%",
-    icon: "water-percent",
-  },
-  {
-    key: "ultraviolet_index_clear_sky",
-    label: "UV Index",
-    unit: "",
-    icon: "weather-sunny-alert",
-  },
-  {
-    key: "wind_from_direction",
-    label: "Wind Dir",
-    unit: "°",
-    icon: "compass",
-  },
-  {
-    key: "wind_speed",
-    label: "Wind Speed",
-    unit: "m/s",
-    icon: "weather-windy",
-  },
-] as const;
-
-export const WeatherDetailsCollapsible: React.FC<WeatherDetailsListProps> = ({
-  details,
-  textSize = "bodyMedium",
-}) => {
-  const [expanded, setExpanded] = useState(false);
-
-  return (
-    <LazyCollapsible
-      title="More Weather Details"
-      icon="widgets"
-      expanded={expanded}
-      onToggle={() => setExpanded((prev) => !prev)}
-      renderContent={() => (
-        <WeatherDetailsList details={details} textSize={textSize} />
-      )}
-      noDataMsg="No additional weather details"
-    />
-  );
-};
-
-export const WeatherDetailsList: React.FC<WeatherDetailsListProps> = ({
-  details,
-  textSize = "bodyMedium",
-}) => (
-  <View>
-    {detailConfig.map(({ key, label, unit, icon }) => (
-      <WeatherDetail
-        key={key}
-        label={label}
-        icon={icon}
-        value={details[key as keyof WeatherInstantDetails]}
-        unit={unit}
-        textSize={textSize}
-      />
-    ))}
-  </View>
-);
-
-type WeatherDetailProps = {
-  label?: string;
-  icon?: string;
-  value?: number | string;
-  unit?: string;
-  fallback?: string;
-  textSize?: "titleSmall" | "titleMedium" | "bodyMedium" | "bodySmall";
-};
-
-export const WeatherDetail: React.FC<WeatherDetailProps> = ({
-  label,
-  icon,
-  value,
-  unit,
-  fallback = "N/A",
-  textSize = "titleSmall",
-}) => {
-  const { colors } = useTheme();
-  return (
-    <View
-      style={{
-        flexDirection: "row",
-        alignItems: "center",
-        gap: 6,
-      }}
-    >
-      {icon && <Icon source={icon} size={16} color={colors.primary} />}
-      <Text variant={textSize}>
-        {label && <Text style={{ fontWeight: "bold" }}>{label}: </Text>}
-        {value !== undefined && value !== null
-          ? `${value}${unit ? ` ${unit}` : ""}`
-          : fallback}
-      </Text>
-    </View>
-  );
-};
-
 // ------ WEATHER HOURS LOGIC ---------
 type WeatherHoursProps = {
   hours: WeatherTimeseriesEntry[];
